apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
httpx==0.27.2
openai==1.54.4
anthropic==0.39.0
google-generativeai==0.8.3
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==43.0.3
